Fixed memory corruption error. shardedVector was int, needed to be long.

+++ MI.h
@@ -11,8 +11,8 @@
 
 using namespace std;
 
-double entropy(const vector<int>& arr){
-    double s = 0;
+float entropy(const vector<int>& arr){
+    float s = 0;
     for (int p: arr){
         if (p!=0){
             s += -p*log(p);
@@ -21,8 +21,8 @@ double entropy(const vector<int>& arr){
     return s;
 }
 
-double mi(const vector<int> &x, const vector<int> &y){
-    double Hx, Hy, Hxy;
+float mi(const vector<int> &x, const vector<int> &y){
+    float Hx, Hy, Hxy;
     auto n_samples = x.size();
 
     Hx = entropy(bincount(x))/n_samples + log(n_samples);
@@ -32,8 +32,8 @@ double mi(const vector<int> &x, const vector<int> &y){
     return Hx + Hy - Hxy;
 }
 
-void miMatrixBlock(const vector<vector<int>>& data, vector<vector<double>>& mi_matrix,
-                   const vector<pair<int, int>>& idx_subset){
+void miMatrixBlock(const vector<vector<int>>& data, vector<vector<float>>& mi_matrix,
+                   const vector<pair<long, long>>& idx_subset){
     int i = 0, j = 0;
 
     for (const auto& idxs : idx_subset){
